feat(vs2022_third): C4018 fix signed/unsigned mismatch in NanoSpec — 5 files (batch 7)
Cast int loop variables to DWORD/unsigned int in comparisons:
- NanoSpec/StressLineSectionSettingDlg.cpp: 30 instances (i,row→DWORD)
- NanoSpec/ChifTransiStress.cpp: 2 instances (i→DWORD in <= comparisons)
- NanoSpec/RecipeThicknessSettingView.cpp: 4 instances (i,j→DWORD)
- NanoSpec/SrVideoCalibrationSettingDlg.cpp: 2 instances (i→unsigned int)
- NanoSpec/StressElasticModulusListDlg.cpp: 2 instances (i→unsigned int)

Total: 40 C4018 warnings fixed (targeting all remaining 40).
No behavior changes, only explicit casts to resolve type mismatch.

+++ SRC/NanoSpec/ChifTransiStress.cpp
@@ -942,7 +942,7 @@ void CChiefTransiStress::enter_ST_STRS_LMVSTAPOS(const cEventParams* const pEvPa
 					break;
 				}
 			}
-			if (l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum <= i) {
+			if (l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum <= (DWORD)i) {
 				TCHAR l_tszLog[128];
 				_stprintf(l_tszLog, _T("No Sections in Line No.%d."), m_dwLineNo);
 				((CChiefView*) m_pcChiefView)->LogChief(l_tszLog);
@@ -1130,7 +1130,7 @@ void CChiefTransiStress::enter_ST_STRS_LMVENDPOS(const cEventParams* const pEvPa
 							break;
 						}
 					}
-					if (l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum <= i) {
+					if (l_StressConfig.Line[m_dwLineNo - 1].dwSectionNum <= (DWORD)i) {
 						TCHAR l_tszLog[128];
 						_stprintf(l_tszLog, _T("No Sections in Line No.%d."), m_dwLineNo);
 						((CChiefView*) m_pcChiefView)->LogChief(l_tszLog);
